NZ0 e_particles, e_room_fg, popup, e_stage_name
