+++ src/st/no3/48A84.c
@@ -317,7 +317,7 @@ void func_801CAD28(s16 sfxId) {
     }
 }
 
-#include "../entity_stage_name_popup.h"
+#include "../e_stage_name.h"
 
 #include "../entity_soul_steal_orb.h"
