+++ src/st/np3/402F4.c
@@ -321,7 +321,7 @@ void func_801C2598(s16 sfxId) {
     }
 }
 
-#include "../entity_stage_name_popup.h"
+#include "../e_stage_name.h"
 
 #include "../entity_soul_steal_orb.h"
