--- src/st/nz0/48ADC.c
@@ -1,8 +1,0 @@
-/*
- * Overlay: NZ0
- * Enemy: Stage Name Popup
- */
-
-#include "nz0.h"
-
-#include "../entity_stage_name_popup.h"
