+++ src/st/nz0/e_room_fg.c
