+++ src/st/nz0/e_stage_name.c
@@ -1,3 +1,3 @@
-#include "wrp.h"
+#include "nz0.h"
 
 #include "../e_stage_name.h"
